backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
langchain==0.0.350
openai==1.3.7
faiss-cpu==1.7.4
sentence-transformers==2.2.2
numpy==1.24.3
scikit-learn==1.3.2
python-multipart==0.0.6
setuptools>=68.0.0
